Move MuteStateController into module
